use a wrapper-style component
